+++ app/src/main/res/layout/swipe_item_box.xml
@@ -1,15 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:padding="16dp"
     tools:context=".SwipeItemBox">
+
 
     <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="16dp"
         android:id="@+id/detail_card"
         android:hapticFeedbackEnabled="true"
         app:cardCornerRadius="6dp"
@@ -20,6 +22,9 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
 
     <RelativeLayout
         android:layout_width="wrap_content"
@@ -51,5 +56,8 @@
             android:textColor="?android:attr/textColorPrimary" />
 
     </RelativeLayout>
-    </com.google.android.material.card.MaterialCardView>
-</ScrollView>
+
+    </ScrollView>
+
+</com.google.android.material.card.MaterialCardView>
+</RelativeLayout>
